buffer: moved to its own package under misc
